Refactor Twilio Chat code into a manager
